+++ frontend/packages/dev-console/package.json
@@ -27,7 +27,8 @@
       "icons": "src/utils/icons.tsx",
       "catalog": "src/components/catalog",
       "fileUpload": "src/components/jar-file-upload/index.ts",
-      "topology": "src/components/topology"
+      "topology": "src/components/topology",
+      "perspective": "src/utils/perspective.tsx"
     }
   }
 }
