+++ frontend/packages/dev-console/src/utils/__tests__/usePerspectiveDetection.spec.ts
@@ -2,7 +2,7 @@
 // @ts-ignore: FIXME missing exports due to out-of-sync @types/react-redux version
 import { useSelector } from 'react-redux';
 import { testHook } from '../../../../../__tests__/utils/hooks-utils';
-import { usePerspectiveDetection } from '../usePerspectiveDetection';
+import { usePerspectiveDetection } from '../perspective';
 
 jest.mock('react-redux', () => ({
   useSelector: jest.fn(),
